오류 수정

+++ src/NcloudMailerDriver.php
@@ -176,7 +176,7 @@
 
 		protected function getTimestamp()
 		{
-			return round(microtime(true) * 1000);
+			return (int)round(microtime(true) * 1000);
 		}
 
 	}
